refactor(BaseBallNumberMaker):  파일 위치 변경
